Move GraphQL types into separate modules (#343)
* Move GraphQL types into separate modules

* Clean up imports

* Minor docstring change

* Rename all_documents to collection

* Add entry to CHANGELOG.md
