+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Restructure `graphql` module [#307](https://github.com/p2panda/aquadoggo/pull/307)
 - Removed replication service for now, preparing for new replication protocol [#296](https://github.com/p2panda/aquadoggo/pull/296)
 - Bring back e2e GraphQL API tests [#342](https://github.com/p2panda/aquadoggo/pull/342)
+- Move GraphQL `types` into separate modules [#343](https://github.com/p2panda/aquadoggo/pull/343)
 
 ### Fixed
