+++ aquadoggo/src/db/stores/mod.rs
@@ -12,4 +12,6 @@ mod schema;
 mod task;
 
 pub use operation::OperationCursor;
-pub use query::{PaginationCursor, Query, QueryResponse, RelationList, RelationListType};
+pub use query::{
+    PaginationCursor, PaginationData, Query, QueryResponse, RelationList, RelationListType,
+};
