+++ aquadoggo/src/graphql/mod.rs
@@ -1,13 +1,16 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 pub mod constants;
+pub mod input_values;
 pub mod mutations;
+pub mod objects;
 pub mod queries;
+pub mod resolvers;
+pub mod responses;
 pub mod scalars;
 mod schema;
 #[cfg(test)]
 mod tests;
-pub mod types;
 pub mod utils;
 
 pub use schema::{build_root_schema, GraphQLSchemaManager};
