+++ aquadoggo/src/graphql/objects/mod.rs
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+mod document;
+mod document_collection;
+mod document_fields;
+mod document_meta;
+
+pub use document::{build_document_object, build_paginated_document_object};
+pub use document_collection::build_document_collection_object;
+pub use document_fields::build_document_fields_object;
+pub use document_meta::DocumentMeta;
