+++ aquadoggo/src/graphql/queries/mod.rs
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-mod all_documents;
+mod collection;
 mod document;
 mod next_args;
 
-pub use all_documents::build_all_documents_query;
+pub use collection::build_collection_query;
 pub use document::build_document_query;
 pub use next_args::build_next_args_query;
