+++ aquadoggo/src/graphql/responses/mod.rs
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+mod next_arguments;
+
+pub use next_arguments::NextArguments;
