+++ aquadoggo/src/graphql/responses/next_arguments.rs
