--- aquadoggo/src/graphql/types/mod.rs
@@ -1,23 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-
-mod document;
-mod document_collection;
-mod document_fields;
-mod document_meta;
-mod filter_input;
-mod filters;
-mod next_arguments;
-mod ordering;
-
-pub use document::{DocumentSchema, DocumentValue, PaginatedDocumentSchema};
-pub use document_collection::{DocumentCollection, PaginationData};
-pub use document_fields::DocumentFields;
-pub use document_meta::DocumentMeta;
-pub use filter_input::{FilterInput, MetaFilterInput};
-pub use filters::{
-    BooleanFilter, DocumentIdFilter, DocumentViewIdFilter, FloatFilter, IntegerFilter, OwnerFilter,
-    PinnedRelationFilter, PinnedRelationListFilter, RelationFilter, RelationListFilter,
-    StringFilter,
-};
-pub use next_arguments::NextArguments;
-pub use ordering::{OrderBy, OrderDirection};
